bug subida de imagenes (arreglado)

+++ models/post.model.js
@@ -3,9 +3,10 @@ import conn from "../database/dbConnection.js";
 import { v4 } from "uuid";
 import { uploadPicture } from "../tools/cloudinary.js";
 import fs from "fs-extra";
+import { log } from "console";
 
 class PostModel {
-  constructor() { }
+  constructor() {}
 
   static async getPosts(req = request, res = response) {
     try {
@@ -96,21 +97,35 @@ class PostModel {
         [post_id, description, location, price, type, user_id]
       ); // Realiza la insercion en la tabla posts
       const { files } = req.files;
-      console.log(files);
-      for (const file of files) {
-        const { tempFilePath } = file;
-        const { secure_url, public_id } = await uploadPicture(tempFilePath);
-        await conn.query(
-          `insert into pictures set 
+      if (Array.isArray(files)) {
+        for (const file of files) {
+          const { tempFilePath } = file;
+          const { secure_url, public_id } = await uploadPicture(tempFilePath);
+          await conn.query(
+            `insert into pictures set 
+                      pic_id = ?, 
+                      url = ?, 
+                      post_id = ?;`,
+            [public_id, secure_url, post_id]
+          );
+          console.log(tempFilePath);
+          fs.unlink(tempFilePath); // elimina los archivos temporales despues de cada insercion
+        }
+        return res.status(201).json({
+          message: "Post has been created sucessfully",
+        });
+      }
+      const { tempFilePath } = files;
+      const { secure_url, public_id } = await uploadPicture(tempFilePath);
+      await conn.query(
+        `insert into pictures set 
                   pic_id = ?, 
                   url = ?, 
                   post_id = ?;`,
-          [public_id, secure_url, post_id]
-        );
-        console.log(tempFilePath);
-        fs.unlink(tempFilePath); // elimina los archivos temporales despues de cada insercion
-      }
-
+        [public_id, secure_url, post_id]
+      );
+      console.log(tempFilePath);
+      fs.unlink(tempFilePath); // elimina los archivos temporales despues de cada insercion
       return res.status(201).json({
         message: "Post has been created sucessfully",
       });
